Handler refactoring
Refactored the handlers.go file in part4 tutorial by adding some useful private functions.

+++ part4/responses.go
@@ -11,6 +11,6 @@ type JsonErrorResponse struct {
 }
 
 type ApiError struct {
-	Status int16  `json:"status"`
+	Status int    `json:"status"`
 	Title  string `json:"title"`
 }
